refactor: rename submissionParams to codeSubmissionParams
